Refactor package structure for better organization of controller and repository classes
